+++ src/snowflake/snowpark/_internal/ast/utils.py
